requirements.txt
google-cloud-texttospeech
opencv-python
mediapipe
gpiozero
pyttsx3
python-dotenv
speechrecognition
pvporcupine
pyaudio
```